% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to_zero.R
\name{na_to_zero}
\alias{na_to_zero}
\title{Turn NAs in a datatable to zero}
\usage{
na_to_zero(DT, excluded = NULL, included = NULL)
}
\arguments{
\item{DT}{a data table passed by reference.}

\item{excluded}{a character vector of column names to exclude -all others included.}

\item{included}{a character vector of column names to include -all others excluded}
}
\value{
No return value.  The data table is changed in place.
}
\description{
This function works on the data.table in place.
}
