% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_extensions.R
\name{na_to_zero}
\alias{na_to_zero}
\title{turn all NAs in a datatable to zero}
\usage{
na_to_zero(DT)
}
\arguments{
\item{DT}{a data table passed by reference.}
}
\value{
No return value.  The data table is changed in place.
}
\description{
This function works on the data.table in place.
}

